refactor: changing id types in repositories to Long

+++ src/main/java/dev/gustavo/math/repository/ChallengeRepository.java
@@ -3,5 +3,5 @@ package dev.gustavo.math.repository;
 import dev.gustavo.math.entity.Challenge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
 }
